Job level Workflows, Score Cards & Forms + Default Schedule + Attachments with name

+++ app/scheduling/schedules/mutations/deleteSchedule.tsx
@@ -1,22 +1,32 @@
 import db from "db"
-import { resolver } from "blitz"
+import { Ctx, resolver } from "blitz"
 import * as z from "zod"
 
-export default resolver.pipe(resolver.zod(z.string()), resolver.authorize(), async (scheduleId) => {
-  const scheduleCalendar = await db.jobUserScheduleCalendar.count({
-    where: { scheduleId },
-  })
+export default resolver.pipe(
+  resolver.zod(z.string()),
+  resolver.authorize(),
+  async (scheduleId, ctx: Ctx) => {
+    const scheduleCalendar = await db.stageUserScheduleCalendar.count({
+      where: { scheduleId },
+    })
+    if (scheduleCalendar > 0) {
+      throw new Error("Can't delete, schedule is being used.")
+    }
 
-  if (scheduleCalendar > 0) {
-    throw new Error("Can't delete, schedule is being used.")
+    const defaultSchedule = await db.defaultSchedule.findFirst({
+      where: { userId: ctx?.session?.userId || "0" },
+    })
+    if (defaultSchedule?.scheduleId === scheduleId) {
+      throw new Error("You cannot delete the default Schedule.")
+    }
+
+    await db.dailySchedule.deleteMany({
+      where: { scheduleId },
+    })
+    await db.schedule.delete({
+      where: { id: scheduleId },
+    })
+
+    return true
   }
-
-  await db.dailySchedule.deleteMany({
-    where: { scheduleId },
-  })
-  await db.schedule.delete({
-    where: { id: scheduleId },
-  })
-
-  return true
-})
+)
